pyproject.toml
[project]
name = "test-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate==0.18.0",
    "audiocaps-download>=1.0.0",
    "audioldm>=0.0.10",
    "audioread>=3.0.0",
    "datasets==2.1.0",
    "diffusers==0.18.2",
    "einops>=0.6.0",
    "h5py==3.8.0",
    "huggingface-hub==0.13.3",
    "importlib-metadata==6.3.0",
    "ipython==8.12.0",
    "librosa>=0.9.2",
    "matplotlib>=3.5.0",
    "mysqlclient>=2.1.0",
    "numpy>=1.22.0",
    "omegaconf>=2.2.0",
    "packaging==23.1",
    "pandas>=1.5.0",
    "pillow>=9.3.0",
    "protobuf==3.20.*",
    "pytorch-lightning>=1.9.0",
    "resampy>=0.4.0",
    "scikit-image==0.19.3",
    "scikit-learn==1.2.2",
    "scipy>=1.9.0",
    "soundfile>=0.11.0",
    "torch>=1.13.1",
    "torchaudio>=0.13.1",
    "torchlibrosa==0.1.0",
    "torchvision==0.14.1",
    "tqdm>=4.64.0",
    "transformers==4.27.0",
    "wandb==0.12.14",
]
